geoanalysis-env/postgres-gis-network/requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bokeh==3.1.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dask==2023.4.0
dask-geopandas==0.3.0
debugpy==1.6.7
decorator==5.1.1
distributed==2023.4.0
et-xmlfile==1.1.0
executing==1.2.0
Fiona==1.9.3
fonttools==4.39.3
fsspec==2023.4.0
geojson==3.0.1
geopandas==0.12.2
greenlet==2.0.2
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.5
munch==2.5.0
nest-asyncio==1.5.6
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
partd==1.4.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.4.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.15
stack-data==0.6.2
tblib==1.7.0
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.0
wcwidth==0.2.6
xyzservices==2023.2.0
zict==3.0.0
zipp==3.15.0